% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsademo.R
\name{baseurlfunc}
\alias{baseurlfunc}
\title{baseurlfunc: Prints the Google Maps API URL}
\usage{
baseurlfunc(x)
}
\arguments{
\item{x}{NULL}
}
\value{
the maps API URL
}
\description{
baseurlfunc: Prints the Google Maps API URL
}
\examples{
baseurl <- baseurlfunc()
}

