% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsademo.R
\name{coffeetime}
\alias{coffeetime}
\title{coffeetime: Calculates the distance to the nearest Starbucks}
\usage{
coffeetime(place)
}
\arguments{
\item{place}{A vector of length 2 containing the numeric lat in position 1,
and the numeric long in position 2, e.g. "41.42326, -73.94992"}
}
\value{
A numeric vector of length 1 containing distance to the nearest Starbucks from \code{place} in kilometers.
}
\description{
coffeetime: Calculates the distance to the nearest Starbucks
}
\examples{
\dontrun{
woods <- c(lat=41.42326, lng=-73.94992)
coffeetime(sb, woods)
}
}

