% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsademo.R
\docType{package}
\name{dsademo}
\alias{dsademo}
\alias{dsademo-package}
\title{\pkg{dsademo}
Provides support information for Derek Darves' teaching audition}
\description{
\pkg{dsademo}
Provides support information for Derek Darves' teaching audition
}

