% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsademo.R
\docType{data}
\name{geo_bucks}
\alias{geo_bucks}
\title{Geo Coordinates for US Starbucks locations}
\format{data.frame with 10,843 rows and two columns:
\describe{
\item{Latitude}{Latitude coordinate}
\item{Longitude}{Longitude coordinate}
}}
\source{
\url{https://opendata.socrata.com/Business/All-Starbucks-Locations-in-the-US-Map/ddym-zvjk}
}
\usage{
geo_bucks
}
\description{
Geo Coordinates for US Starbucks locations
}
\examples{
  geo_bucks
}
\keyword{datasets}

