% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsademo.R
\name{get_sb}
\alias{get_sb}
\title{Load the Starbucks geo data}
\source{
\href{https://opendata.socrata.com/Business/All-Starbucks-Locations-in-the-US-Map/ddym-zvjk}{socrata}
}
\usage{
get_sb()
}
\value{
A data.frame
}
\description{
Load the Starbucks geo data
}

