% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsademo.R
\name{rcurlbuild}
\alias{rcurlbuild}
\title{rcurlbuild: Sets the environment for dsademo}
\usage{
rcurlbuild(x)
}
\arguments{
\item{x}{a character vector of json text}
}
\value{
A formatted object
}
\description{
rcurlbuild: Sets the environment for dsademo
}

