% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_bucks.R
\docType{data}
\name{geo_bucks}
\alias{geo_bucks}
\title{Geo Coordinates for US Starbucks locations}
\format{Data frame with two columns.}
\source{
\url{https://opendata.socrata.com/Business/All-Starbucks-Locations-in-the-US-Map/ddym-zvjk}
}
\usage{
geo_bucks
}
\description{
#' Load SB geocodes
}
\examples{
  geo_bucks
}
\keyword{datasets}

